requirements.txt
paho-mqtt
asyncua
